requirements.txt
streamlit>=1.32.0
google-genai>=0.3.0
groq>=0.5.0
plotly>=5.18.0
pdfplumber>=0.10.0
python-docx>=1.1.0
duckduckgo-search>=5.0.0
python-dotenv>=1.0.0
fpdf2>=2.7.0